requirements.txt
anyio==4.4.0
click==8.1.7
exceptiongroup==1.2.2
graphql-core==3.2.3
h11==0.14.0
idna==3.7
libcst==1.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2==2.9.9
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
ruff==0.5.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
strawberry-graphql==0.237.3
typer==0.12.3
typing_extensions==4.12.2
uvicorn==0.30.5
